Filter down to printed works
Do this (for now) before dumping subjects and concepts.

We have to do the filtering in Python. We could have used JSONPath
if works were a single JSON object, but the JSONL format means that
each work appears to JSONPath as a separate JSON object, preventing
use of JSONPath filters, which work only on arrays.